requirements.txt
Flask
Flask-SQLAlchemy
matplotlib
numpy
pandas
psycopg2-binary
python-dotenv
requests
streamlit